+++ src/main/drivers/bus.h
@@ -64,7 +64,7 @@ typedef struct busDevice_s {
     uint8_t deviceCount;
     dmaChannelDescriptor_t *dmaTx;
     dmaChannelDescriptor_t *dmaRx;
-#ifndef UNIT_TEST
+#ifdef USE_DMA
     // Use a reference here as this saves RAM for unused descriptors
 #if defined(USE_FULL_LL_DRIVER)
     LL_DMA_InitTypeDef          *initTx;
@@ -73,7 +73,7 @@ typedef struct busDevice_s {
     DMA_InitTypeDef             *initTx;
     DMA_InitTypeDef             *initRx;
 #endif
-#endif // UNIT_TEST
+#endif // USE_DMA
     volatile struct busSegment_s* volatile curSegment;
     bool initSegment;
 } busDevice_t;
@@ -94,7 +94,7 @@ typedef struct extDevice_s {
             uint8_t address;
         } mpuSlave;
     } busType_u;
-#ifndef UNIT_TEST
+#ifdef USE_DMA
     // Cache the init structure for the next DMA transfer to reduce inter-segment delay
 #if defined(USE_FULL_LL_DRIVER)
     LL_DMA_InitTypeDef          initTx;
@@ -103,7 +103,7 @@ typedef struct extDevice_s {
     DMA_InitTypeDef             initTx;
     DMA_InitTypeDef             initRx;
 #endif
-#endif // UNIT_TEST
+#endif // USE_DMA
     // Support disabling DMA on a per device basis
     bool useDMA;
     // Per device buffer reference if needed
