+++ src/main/drivers/system.h
@@ -71,19 +71,14 @@ int32_t clockCyclesTo10thMicros(int32_t clockCycles);
 int32_t clockCyclesTo100thMicros(int32_t clockCycles);
 uint32_t clockMicrosToCycles(uint32_t micros);
 uint32_t getCycleCounter(void);
-#if defined(STM32H7) || defined(STM32G4)
 void systemProcessResetReason(void);
-#endif
 
 // memory
 void memoryMappedModeInit(void);
 bool isMemoryMappedModeEnabledOnBoot(void);
 
 void initialiseMemorySections(void);
-#ifdef STM32H7
-void initialiseD2MemorySections(void);
 void systemResetWithoutDisablingCaches(void);
-#endif
 
 void enableGPIOPowerUsageAndNoiseReductions(void);
 // current crystal frequency - 8 or 12MHz
