+++ src/platform/STM32/mk/STM32F7.mk
@@ -141,6 +141,7 @@ VCP_SRC = \
             drivers/usb_io.c
 
 MCU_COMMON_SRC = \
+            stm32/system.c \
             drivers/accgyro/accgyro_mpu.c \
             drivers/bus_i2c_timing.c \
             drivers/dshot_bitbang_decode.c \
@@ -184,6 +185,9 @@ MSC_SRC = \
             msc/emfat_file.c \
             msc/usbd_storage_sdio.c \
             msc/usbd_storage_sd_spi.c
+
+SPEED_OPTIMISED_SRC := $(SPEED_OPTIMISED_SRC) \
+            stm32/system.c
 
 DSP_LIB := $(ROOT)/lib/main/CMSIS/DSP
 DEVICE_FLAGS += -DARM_MATH_MATRIX_CHECK -DARM_MATH_ROUNDING -D__FPU_PRESENT=1 -DUNALIGNED_SUPPORT_DISABLE -DARM_MATH_CM7
