+++ src/platform/STM32/mk/STM32H5.mk
@@ -149,6 +149,7 @@ VCP_SRC =
             drivers/usb_io.c
 
 MCU_COMMON_SRC = \
+            stm32/system.c \
             drivers/bus_i2c_timing.c \
             drivers/bus_quadspi.c \
             drivers/dshot_bitbang_decode.c \
@@ -196,6 +197,9 @@ MSC_SRC =
             msc/emfat_file.c \
             msc/usbd_storage_sd_spi.c \
             msc/usbd_storage_sdio.c
+
+SPEED_OPTIMISED_SRC := $(SPEED_OPTIMISED_SRC) \
+            stm32/system.c
 
 DSP_LIB := $(ROOT)/lib/main/CMSIS/DSP
 DEVICE_FLAGS += -DARM_MATH_MATRIX_CHECK -DARM_MATH_ROUNDING -DUNALIGNED_SUPPORT_DISABLE -DARM_MATH_CM7
